cache: move package to plumbing
Because cache package is only intended to be used at internal level, we move it to the plumbing package.
