+++ plumbing/cache/object.go
